🏗️ Architect: Modularized apts/cache.py
Decomposed the monolithic `apts/cache.py` God Object into a specialized package structure:
- `apts/cache/skyfield.py`: Timescale and ephemeris caching.
- `apts/cache/catalogs.py`: Hipparcos and MPC orbit data caching.
- `apts/cache/nasa.py`: NASA Comet data caching.
- `apts/cache/scoring.py`: Internal scoring cache management.
- `apts/cache/base.py`: Global lifecycle functions (`download_all_data`, `clear_cache`).

Maintained backward compatibility for the public API via `apts/cache/__init__.py`. Updated unit test patches in `tests/unit/test_cache.py` to target the new submodule locations.
